Error log when config can't be loaded
I've shot myself in the food too many times developing it so people in
prod will do it too. Sadly we can't just info log it because error is
the default log level for the root logger. While we could change that we
also read the logging config from the config file so...